requirements.txt
Flask==1.1.1
spanish_sentiment_analysis==1.0.0
pandas==0.24.2
Werkzeug==0.15.4
psycopg2-binary==2.8.4
requests==2.22.0
SQLAlchemy==1.3.5
classifier==2.0
flask_cors==3.0.8
grequests==0.4.0
jsonify==0.5
langdetect==1.0.7
pyquery==1.4.1
textblob==0.15.3
Unidecode==1.1.1
IPython==7.9.0
